requirements.txt
asgiref==3.8.1
astroid==3.3.5
autopep8==2.3.1
blinker==1.9.0
click==8.1.8
dill==0.3.9
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.1.0
gunicorn==23.0.0
importlib_metadata==8.5.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
PyJWT==2.9.0
pylint==3.3.1
python-dotenv==1.0.1
sqlparse==0.5.2
tomli==2.1.0
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
